fix : user 기본키 자료형에 맞게 변경

+++ src/main/java/uos/cineseoul/utils/CustomUserDetails.java
@@ -21,7 +21,7 @@ public class CustomUserDetails implements UserDetails {
 
     private static final long serialVersionUID = 1L;
 
-    private int num;
+    private Long num;
 
     private String name;
